% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbx.R
\name{dbxConnect}
\alias{dbxConnect}
\title{Create a database connection}
\usage{
dbxConnect(adapter = NULL, url = NULL, ...)
}
\arguments{
\item{adapter}{The database adapter to use}

\item{url}{A database URL}

\item{...}{Arguments to pass to dbConnect}
}
\description{
Create a database connection
}
\examples{
# Postgres
db <- dbxConnect(adapter="postgres", dbname="mydb")

# MySQL
db <- dbxConnect(adapter="mysql", dbname="mydb")

# SQLite
db <- dbxConnect(adapter="sqlite", dbname=":memory:")

db <- dbxConnect(adapter=odbc(), database="mydb")
}
