% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbx.R
\name{dbxDelete}
\alias{dbxDelete}
\title{Delete records}
\usage{
dbxDelete(conn, table, where = NULL)
}
\arguments{
\item{conn}{A DBIConnection object}

\item{table}{The table name to delete records from}

\item{where}{A data frame of records to delete}
}
\description{
Delete records
}
