% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbx.R
\name{dbxDisconnect}
\alias{dbxDisconnect}
\title{Close a database connection}
\usage{
dbxDisconnect(conn)
}
\arguments{
\item{conn}{A DBIConnection object}
}
\description{
Close a database connection
}
\examples{
\dontrun{

dbxDisconnect(db)
}
}
