% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbx.R
\name{dbxInsert}
\alias{dbxInsert}
\title{Insert records}
\usage{
dbxInsert(conn, table, records)
}
\arguments{
\item{conn}{A DBIConnection object}

\item{table}{The table name to insert}

\item{records}{A data frame of records to insert}
}
\description{
Insert records
}
