% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{dbxSelect}
\alias{dbxSelect}
\title{Select records}
\usage{
dbxSelect(conn, statement, retries = 0)
}
\arguments{
\item{conn}{A DBIConnection object}

\item{statement}{The SQL statement to use}

\item{retries}{Max retries if query fails}
}
\description{
Select records
}
\examples{
db <- dbxConnect(adapter="sqlite", dbname=":memory:")
DBI::dbCreateTable(db, "forecasts", data.frame(id=1:3, temperature=20:22))

records <- dbxSelect(db, "SELECT * FROM forecasts")
}
