% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbx.R
\name{dbxUpsert}
\alias{dbxUpsert}
\title{Upsert records}
\usage{
dbxUpsert(conn, table, records, where_cols, batch_size = NULL)
}
\arguments{
\item{conn}{A DBIConnection object}

\item{table}{The table name to upsert}

\item{records}{A data frame of records to upsert}

\item{where_cols}{The columns to use for WHERE clause}

\item{batch_size}{The number of records to upsert in a single transaction (defaults to all)}
}
\description{
Upsert records
}
