% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbx.R
\name{dbxConnect}
\alias{dbxConnect}
\title{Create a database connection}
\usage{
dbxConnect(url = NULL, adapter = NULL, ...)
}
\arguments{
\item{url}{A database URL}

\item{adapter}{The database adapter to use}

\item{...}{Arguments to pass to dbConnect}
}
\description{
Create a database connection
}
