% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbx.R
\name{dbxDelete}
\alias{dbxDelete}
\title{Delete records}
\usage{
dbxDelete(conn, table, where = NULL, batch_size = NULL)
}
\arguments{
\item{conn}{A DBIConnection object}

\item{table}{The table name to delete records from}

\item{where}{A data frame of records to delete}

\item{batch_size}{The number of records to delete in a single statement (defaults to all)}
}
\description{
Delete records
}
\examples{
bad_records <- data.frame(id=c(1, 2))
dbxDelete(con, table, where=bad_records)

dbxDelete(con, table)
}
