% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbx.R
\name{dbxInsert}
\alias{dbxInsert}
\title{Insert records}
\usage{
dbxInsert(conn, table, records, batch_size = NULL)
}
\arguments{
\item{conn}{A DBIConnection object}

\item{table}{The table name to insert}

\item{records}{A data frame of records to insert}

\item{batch_size}{The number of records to insert in a single statement (defaults to all)}
}
\description{
Insert records
}
