% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbx.R
\name{dbxSelect}
\alias{dbxSelect}
\title{Select records}
\usage{
dbxSelect(conn, statement)
}
\arguments{
\item{conn}{A DBIConnection object}

\item{statement}{The SQL statement to use}
}
\description{
Select records
}
\examples{
\donttest{
db <- dbxConnect(adapter="sqlite", dbname=":memory:")

records <- dbxSelect(db, "SELECT * FROM forecasts")

}
}
