% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbx.R
\name{dbxUpdate}
\alias{dbxUpdate}
\title{Update records}
\usage{
dbxUpdate(conn, table, records, where_cols)
}
\arguments{
\item{conn}{A DBIConnection object}

\item{table}{The table name to update}

\item{records}{A data frame of records to insert}

\item{where_cols}{The columns to use for WHERE clause}
}
\description{
Update records
}
